移植批次7 第2项：Java 尾巴（distinctive/地图名/掉落查询/倍率券 float 化）
MonsterDropEntry+MonsterInformationProvider 接 distinctive 列；MapFactory 补
7亿段→chinese（singapore→SG 不搬，BeiDou String.wz 无该节点）；@whatdropsfrom
吸收物品图标富文本、保留按怪名搜索；@whodrops already-fixed（批次5脚本中心）。
倍率券全链 float 化（NxcouponsDO/Server/Character/ExpLogger + V1000.1.7 改列
并插 1.5 倍经验/掉落券）。gachapon 16 文件判定为死代码路径，与 doGachapon(quantity)
及 4 个 gacha 脚本合并成工作包挪第 3 项脚本组。

+++ gms-server/src/main/java/org/gms/net/server/Server.java
@@ -102,7 +102,7 @@ public class Server {
     }
 
     private static final Set<Integer> activeFly = new HashSet<>();
-    private static final Map<Integer, Integer> couponRates = new HashMap<>(30);
+    private static final Map<Integer, Float> couponRates = new HashMap<>(30);
     private static final List<Integer> activeCoupons = new LinkedList<>();
     private ChannelDependencies channelDependencies;
 
@@ -557,7 +557,7 @@ public class Server {
         return Math.max(0, nextDay.getTimeInMillis() - System.currentTimeMillis());
     }
 
-    public Map<Integer, Integer> getCouponRates() {
+    public Map<Integer, Float> getCouponRates() {
         return couponRates;
     }
